extracted interface of the PluginManager
